IN PROGRESS: AllTests: 0 failures, 0 errors.

+++ src/com/amazon/ion/impl/AbstractBuffer.java
@@ -20,52 +20,52 @@ abstract class AbstractBuffer {
     /**
      * The index of the next byte in the buffer that is available to be read. Always less than or equal to `limit`.
      */
-    int offset = 0;
+    long offset = 0;
 
     /**
      * The index at which the next byte received will be written. Always greater than or equal to `offset`.
      */
-    int limit = 0;
+    long limit = 0;
 
-    int capacity;
+    long capacity;
 
     ByteBuffer byteBuffer;
 
     State state = State.READY;
 
-    int bytesRequested = 0;
+    long bytesRequested = 0;
 
-    abstract int peek(int index);
+    abstract int peek(long index);
 
     // TODO abstraction?
-    ByteBuffer getByteBuffer(int startIndex, int endIndex) {
+    ByteBuffer getByteBuffer(long startIndex, long endIndex) {
         // Setting the limit to the capacity first is required because setting the position will fail if the new
         // position is outside the limit.
-        byteBuffer.limit(capacity);
-        byteBuffer.position(startIndex);
-        byteBuffer.limit(endIndex);
+        byteBuffer.limit((int) capacity);
+        byteBuffer.position((int) startIndex);
+        byteBuffer.limit((int) endIndex);
         return byteBuffer;
     }
 
-    int available() {
+    long available() {
         return availableAt(offset);
     }
 
-    int availableAt(int index) {
+    long availableAt(long index) {
         return limit - index;
     }
 
-    abstract void copyBytes(int position, byte[] destination, int destinationOffset, int length);
+    abstract void copyBytes(long position, byte[] destination, int destinationOffset, int length);
 
-    boolean fill(int numberOfBytes) throws Exception {
+    boolean fill(long numberOfBytes) throws Exception {
         return fillAt(offset, numberOfBytes);
     }
 
-    abstract boolean fillAt(int index, int numberOfBytes) throws Exception;
+    abstract boolean fillAt(long index, long numberOfBytes) throws Exception;
 
-    abstract boolean seek(int numberOfBytes) throws IOException;
+    abstract boolean seek(long numberOfBytes) throws IOException;
 
-    boolean seekTo(int index) throws IOException {
+    boolean seekTo(long index) throws IOException {
         return seek(index - offset);
     }
 
@@ -84,7 +84,7 @@ abstract class AbstractBuffer {
         }
     }
 
-    int getOffset() {
+    long getOffset() {
         return offset;
     }
